Swift: split SealedSender certificates out to their own file
